mastro_consumer/pyproject.toml
[tool.poetry]
name = "mastro-consumer"
version = "0.1.0"
description = ""
authors = ["Alvaro Garcia Molino <alvaro.garcia.molino@gmolino.com>"]
readme = "README.md"
packages = [{include = "mastro_consumer"}]

[tool.poetry.dependencies]
python = "^3.10"
pika = "^1.3.1"
rich = "^13.3.2"
psycopg2-binary = "^2.9.5"
pydantic = "^2.6.4"
pydantic-settings = "^2.2.1"
pendulum = "^3.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
